+++ money-pilot/src/contexts/SetupContext.tsx
@@ -77,6 +77,14 @@ export const SetupProvider: React.FC<{ children: ReactNode }> = ({
       // 3. If setup was completed, never show again regardless of user age
       const shouldShowSetup = userIsNew && !setupWasCompleted;
 
+      console.log("🔍 Setup Debug:", {
+        userId: user.uid,
+        userIsNew,
+        setupWasCompleted,
+        shouldShowSetup,
+        savedProgress: savedProgress ? "exists" : "null",
+      });
+
       setIsFirstTimeUser(shouldShowSetup);
       setIsLoading(false);
     } catch (error) {
