+++ money-pilot/src/services/sharedFinanceDataSync.ts
@@ -274,44 +274,3 @@ export const getUserGroupSharingSettings = async (
     return null;
   }
 };
-
-/**
- * Removes a user's shared data from a group
- */
-export const removeUserFromGroup = async (
-  userId: string,
-  groupId: string
-): Promise<void> => {
-  try {
-    const sharedDataRef = ref(
-      db,
-      `sharedFinanceData/${groupId}/members/${userId}`
-    );
-    await remove(sharedDataRef);
-  } catch (error: any) {
-    if (handlePermissionError(error, "skipping user removal from group")) {
-      return;
-    }
-
-    console.error("❌ Error removing user from group:", error);
-    throw error;
-  }
-};
-
-/**
- * Removes all shared data for a group (when group is deleted)
- */
-export const removeGroupSharedData = async (groupId: string): Promise<void> => {
-  try {
-    const sharedDataRef = ref(db, `sharedFinanceData/${groupId}`);
-    await remove(sharedDataRef);
-    console.log("✅ Removed all shared data for group:", groupId);
-  } catch (error: any) {
-    if (handlePermissionError(error, "skipping group shared data removal")) {
-      return;
-    }
-
-    console.error("❌ Error removing group shared data:", error);
-    throw error;
-  }
-};
